year2020/src/test/java/dev/linl33/adventofcode/year2020/test/Day23Test.java
package dev.linl33.adventofcode.year2020.test;

import dev.linl33.adventofcode.lib.function.ThrowingBiFunction;
import dev.linl33.adventofcode.lib.solution.AdventSolution;
import dev.linl33.adventofcode.lib.solution.ResourceIdentifier;
import dev.linl33.adventofcode.testlib.AdventSolutionTest;
import dev.linl33.adventofcode.testlib.Part1Source;
import dev.linl33.adventofcode.year2020.Day23;
import org.junit.jupiter.params.ParameterizedTest;

import java.io.BufferedReader;
import java.util.Map;

import static org.junit.jupiter.api.Assertions.assertEquals;

class Day23Test implements AdventSolutionTest<Integer, Long> {
  @Override
  public AdventSolution<Integer, Long> newSolutionInstance() {
    return new Day23();
  }

  @Override
  public Map<Object, Integer> getPart1Cases() {
    return Map.of(
        newSolutionInstance().getPart1Resource(), 58427369,
        "day23test1", 67384529
    );
  }

  @Override
  public Map<Object, Long> getPart2Cases() {
    return Map.of(
        newSolutionInstance().getPart1Resource(), 111057672960L,
        "day23test1", 149245887792L
    );
  }

  @ParameterizedTest
  @Part1Source
  void testPart1ByLinkedList(ResourceIdentifier resource, int expected, Day23 instance) {
    assertEquals(expected, instance.run((ThrowingBiFunction<Day23, BufferedReader, Integer>) Day23::part1ByLinkedList, resource));
  }
}
